chore: refactored.
moving brml and xy parsing routines in seperate files
