+++ src/__tests__/parseBRML.test.js
@@ -1,7 +1,7 @@
 import { readFileSync } from 'fs';
 import { join } from 'path';
 
-import { parseDiffractogram, parseXY } from '../parser/parser';
+import { parseDiffractogram } from '../parser/parserBRML';
 
 describe('parseBRML', () => {
   const xml = readFileSync(join(__dirname, '../../data/RawData0.xml'), 'utf8');
@@ -26,22 +26,5 @@ describe('parseBRML', () => {
     expect(result.data.y.slice(0, 3)).toStrictEqual([1080, 1024, 1009]);
     expect(result.data.x[3714 - 1]).toStrictEqual(40.0099);
     expect(result.data.y[3714 - 1]).toStrictEqual(472);
-  });
-});
-
-describe('parseXY', () => {
-  const xy = readFileSync(
-    join(__dirname, '../../data/MG1-Cu2O-28_bg_subtracted.xy'),
-    'utf8',
-  );
-
-  it('check the dictionary', () => {
-    const diffractogram = parseXY(xy);
-    expect(diffractogram.meta.userName).toBe('Lab Manager');
-    expect(diffractogram.data.x[0]).toStrictEqual(10);
-    expect(diffractogram.data.y[1]).toStrictEqual(-0.0755227112146954);
-    expect(diffractogram.info.origin).toBe(
-      'Data converted from xy using convert-to-jcamp',
-    );
   });
 });
