+++ src/reader/readXY.js
@@ -1,0 +1,6 @@
+import { parseXY } from '../parser/parseXY';
+
+export function readXY(file) {
+  const diffractogram = parseXY(file);
+  return diffractogram;
+}
